requirements.txt
aiogram==3.20.0.post0
aiohttp==3.9.3
python-dotenv==1.1.0
pytest==8.4.0
pytest-asyncio==1